requirements.txt
Flask==3.0.0
PyGithub==2.1.1
python-dotenv==1.0.0
PyYAML==6.0.1
fuzzywuzzy==0.18.0
python-Levenshtein==0.23.0
requests==2.31.0
gunicorn==21.2.0

# Testing dependencies
pytest==7.4.3
pytest-flask==1.3.0
pytest-cov==4.1.0
pytest-mock==3.12.0